refactor(engine): move legacy writer into StreamProcessor package
