+++ engine/src/main/java/io/camunda/zeebe/streamprocessor/DirectTaskResultBuilder.java
@@ -9,7 +9,6 @@ package io.camunda.zeebe.streamprocessor;
 
 import io.camunda.zeebe.engine.api.TaskResult;
 import io.camunda.zeebe.engine.api.TaskResultBuilder;
-import io.camunda.zeebe.engine.processing.streamprocessor.writers.LegacyTypedStreamWriter;
 import io.camunda.zeebe.protocol.record.RecordType;
 import io.camunda.zeebe.protocol.record.RecordValue;
 import io.camunda.zeebe.protocol.record.RejectionType;
